extract panel buttons logic, readyToTrain and readyForResults -> resultsAvailable
