+++ src/redux.js
@@ -1,4 +1,7 @@
-import { isPanelEnabled, isPanelAvailable } from "./navigationValidation.js";
+import {
+  uniqLabelFeaturesSelected,
+  prevNextButtons
+} from "./navigationValidation.js";
 
 import {
   ColumnTypes,
@@ -1061,86 +1064,12 @@ export function getPredictAvailable(state) {
   );
 }
 
-function isAccuracyAcceptable(state) {
-  const mode = state.mode;
-
-  if (
-    mode &&
-    mode.requireAccuracy &&
-    mode.requireAccuracy > state.historicResults[0].accuracy
-  ) {
-    return false;
-  }
-
-  return true;
+export function getPanelButtons(state) {
+  return prevNextButtons(state);
 }
 
-// Given the current panel, return the appropriate previous & next buttons.
-export function getPanelButtons(state) {
-  let prev, next;
-
-  if (state.currentPanel === "selectDataset") {
-    prev = null;
-    next = isPanelAvailable(state, "dataDisplayLabel")
-      ? { panel: "dataDisplayLabel", text: "Continue" }
-      : isPanelAvailable(state, "dataDisplayFeatures")
-      ? { panel: "dataDisplayFeatures", text: "Continue" }
-      : null;
-  } else if (state.currentPanel === "dataDisplayLabel") {
-    prev = isPanelAvailable(state, "selectDataset")
-      ? { panel: "selectDataset", text: "Back" }
-      : null;
-    next = isPanelAvailable(state, "dataDisplayFeatures")
-      ? { panel: "dataDisplayFeatures", text: "Continue" }
-      : null;
-  } else if (state.currentPanel === "dataDisplayFeatures") {
-    prev = isPanelAvailable(state, "dataDisplayLabel")
-      ? { panel: "dataDisplayLabel", text: "Back" }
-      : null;
-    next = isPanelAvailable(state, "trainModel")
-      ? { panel: "trainModel", text: "Train" }
-      : null;
-  } else if (state.currentPanel === "trainModel") {
-    if (state.modelSize) {
-      prev = null;
-      next = { panel: "generateResults", text: "Continue" };
-    }
-  } else if (state.currentPanel === "generateResults") {
-    if (state.modelSize) {
-      prev = null;
-      next = { panel: "results", text: "Continue" };
-    }
-  } else if (state.currentPanel === "results") {
-    prev = isPanelAvailable(state, "dataDisplayFeatures")
-      ? { panel: "dataDisplayFeatures", text: "Try again" }
-      : null;
-    next = !isAccuracyAcceptable(state)
-      ? null
-      : isPanelAvailable(state, "saveModel")
-      ? { panel: "saveModel", text: "Continue" }
-      : { panel: "continue", text: "Finish" };
-  } else if (state.currentPanel === "saveModel") {
-    prev = { panel: "results", text: "Back" };
-    next = isPanelAvailable(state, "modelSummary")
-      ? { panel: "modelSummary", text: "Save" }
-      : null;
-  } else if (state.currentPanel === "modelSummary") {
-    prev = isPanelAvailable(state, "saveModel")
-      ? { panel: "saveModel", text: "Back" }
-      : null;
-    next = isPanelAvailable(state, "finish")
-      ? { panel: "finish", text: "Finish" }
-      : null;
-  }
-
-  if (prev) {
-    prev.enabled = isPanelEnabled(state, prev.panel);
-  }
-  if (next) {
-    next.enabled = isPanelEnabled(state, next.panel);
-  }
-
-  return { prev, next };
+export function readyToTrain(state) {
+  return uniqLabelFeaturesSelected(state);
 }
 
 /* Returns an object with information for the CrossTab UI.
